+++ src/App/Fossa/Container/Scan.hs
@@ -11,6 +11,7 @@ module App.Fossa.Container.Scan (
 ) where
 
 import App.Fossa.Config.Analyze (WithoutDefaultFilters)
+import App.Fossa.Config.Container.Analyze (GoBinaryDiscovery)
 import App.Fossa.Config.Container.Common (ImageText (unImageText))
 import App.Fossa.Container.Sources.DockerArchive (analyzeFromDockerArchive, revisionFromDockerArchive)
 import App.Fossa.Container.Sources.DockerEngine (analyzeFromDockerEngine, revisionFromDockerEngine)
@@ -89,11 +90,12 @@ scanImage ::
   AllFilters ->
   Flag WithoutDefaultFilters ->
   Bool ->
+  Flag GoBinaryDiscovery ->
   ImageText ->
   Text ->
   Text ->
   m ContainerScan
-scanImage filters withoutDefaultFilters systemDepsOnly imgText dockerHost imageArch = do
+scanImage filters withoutDefaultFilters systemDepsOnly goBinaryDiscovery imgText dockerHost imageArch = do
   parsedSource <- runDockerEngineApi dockerHost $ parseContainerImageSource (unImageText imgText) imageArch
   circePoweredScan <- withSystemTempDir "fossa-container-export-tmp" $ \dir -> do
     tarball <- runWithCirceReexport imgText dir
@@ -101,19 +103,19 @@ scanImage filters withoutDefaultFilters systemDepsOnly imgText dockerHost imageA
     traverse (correctCirceSource parsedSource) scan
   maybe (legacyScan parsedSource) pure circePoweredScan
   where
-    analyzeTarball = context "Analyzing docker archive" . analyzeFromDockerArchive systemDepsOnly filters withoutDefaultFilters
+    analyzeTarball = context "Analyzing docker archive" . analyzeFromDockerArchive systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters
     legacyScan src = do
       case src of
         DockerArchive tarball -> context "Analyzing tarball" $ analyzeTarball tarball
         DockerEngine imgTag ->
           context "Analyzing via Docker engine API" $
-            analyzeFromDockerEngine systemDepsOnly filters withoutDefaultFilters dockerHost imgTag
+            analyzeFromDockerEngine systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters dockerHost imgTag
         Podman img ->
           context "Analyzing via podman" $
-            analyzeFromPodman systemDepsOnly filters withoutDefaultFilters img
+            analyzeFromPodman systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters img
         Registry registrySrc ->
           context "Analyzing via registry" $
-            analyzeFromRegistry systemDepsOnly filters withoutDefaultFilters registrySrc
+            analyzeFromRegistry systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters registrySrc
 
 scanImageNoAnalysis ::
   ( Has Diagnostics sig m